# docs/swagger.yaml
definitions:
...
  main.TicketStatus:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  main.User:
    properties:
      birth_date:
        description: формат "YYYY-MM-DD"
        type: string
      email:
        type: string
      id:
        type: string
      is_blocked:
        type: boolean
      name:
        type: string
    type: object
  main.UserLogin:
    properties:
      email:
        type: string
      password-hash:
        type: string
    type: object
  main.UserRegister:
    properties:
      birth_date:
        description: формат "YYYY-MM-DD"
        type: string
      email:
        type: string
      name:
        type: string
      password-hash:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample
  title: Your API Title
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Логин пользователя с именем пользователя и паролем
      parameters:
      - description: Пользователь
        in: body
        name: user_login
        required: true
        schema:
          $ref: '#/definitions/main.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Токен авторизации
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "401":
          description: Неверное имя пользователя или пароль
          schema:
            type: string
        "500":
          description: Ошибка генерации токена
          schema:
            type: string
      summary: Логин пользователя
...
  /register:
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя с именем пользователя и паролем
      parameters:
      - description: Пользователь
        in: body
        name: user_register
        required: true
        schema:
          $ref: '#/definitions/main.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь зарегистрирован
          schema:
            type: string
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сохранения пользователя
          schema:
            type: string
      summary: Регистрация нового пользователя
...
  /ticket-statuses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.TicketStatus'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить все статусы билетов
      tags:
      - ticket-statuses
    post:
      consumes:
      - application/json
      parameters:
      - description: Статус билета
        in: body
        name: ticket_status
        required: true
        schema:
          $ref: '#/definitions/main.TicketStatus'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.TicketStatus'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать статус билета
      tags:
      - ticket-statuses
  /ticket-statuses/{id}:
    delete:
      parameters:
      - description: ID статуса билета
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Статус не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить статус билета
      tags:
      - ticket-statuses
    get:
      parameters:
      - description: ID статуса билета
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.TicketStatus'
        "404":
          description: Статус билета не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить статус билета по ID
      tags:
      - ticket-statuses
    put:
      consumes:
      - application/json
      parameters:
      - description: ID статуса билета
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные статуса
        in: body
        name: ticket_status
        required: true
        schema:
          $ref: '#/definitions/main.TicketStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.TicketStatus'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Статус не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить статус билета
      tags:
      - ticket-statuses
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.User'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить всех пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Пользователь
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Неверный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Удалено
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить пользователя
      tags:
      - users
    get:
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить пользователя по ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Неверный JSON
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить пользователя
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"